Extract JavaInstallationRegistry interface from implementation
