+++ testing/architecture-test/src/changes/archunit_store/internal-api-nullability.txt
@@ -2091,9 +2091,6 @@ Class <org.gradle.internal.jvm.inspection.DefaultJvmMetadataDetector> is not ann
 Class <org.gradle.internal.jvm.inspection.DefaultJvmVersionDetector> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultJvmVersionDetector.java:0)
 Class <org.gradle.internal.jvm.inspection.InvalidInstallationWarningReporter> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (InvalidInstallationWarningReporter.java:0)
 Class <org.gradle.internal.jvm.inspection.InvalidJvmInstallationCacheInvalidator> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (InvalidJvmInstallationCacheInvalidator.java:0)
-Class <org.gradle.internal.jvm.inspection.JavaInstallationRegistry$Installations> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JavaInstallationRegistry.java:0)
-Class <org.gradle.internal.jvm.inspection.JavaInstallationRegistry$ToolchainDetectionBuildOperation> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JavaInstallationRegistry.java:0)
-Class <org.gradle.internal.jvm.inspection.JavaInstallationRegistry> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JavaInstallationRegistry.java:0)
 Class <org.gradle.internal.jvm.inspection.JvmInstallationMetadata$DefaultJvmInstallationMetadata> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JvmInstallationMetadata.java:0)
 Class <org.gradle.internal.jvm.inspection.JvmInstallationMetadata$FailureInstallationMetadata> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JvmInstallationMetadata.java:0)
 Class <org.gradle.internal.jvm.inspection.JvmInstallationMetadata$JavaInstallationCapability> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (JvmInstallationMetadata.java:0)
